Restructure public item hierarchy
Resolves #44.
