+++ examples/gravity.rs
@@ -7,8 +7,7 @@ use std::rand::thread_rng;
 use std::rand::distributions::{Range, IndependentSample};
 use std::num::Float;
 use nalgebra::{Pnt3, Vec3, FloatPnt, Norm, Orig, zero};
-use acacia::ntree::Tree;
-use acacia::tree::{DataQuery, Node, AssociatedData, Position};
+use acacia::{Tree, DataQuery, Node, AssociatedData, Position};
 use acacia::partition::Ncube;
 
 /// Point mass
